+++ tests/tests.py
@@ -10,7 +10,9 @@ from django.urls import reverse
 from django.utils import timezone
 
 from .models import Article
+from .views import to_page_key
 from infinite_scroll_pagination.paginator import SeekPaginator
+from infinite_scroll_pagination import paginator as inf_paginator
 
 
 class PaginatorTest(TestCase):
@@ -22,83 +24,105 @@ class PaginatorTest(TestCase):
             seconds = datetime.timedelta(seconds=i)
             Article.objects.create(title="%s" % i, date=date, date_unique=date + seconds)
 
-        self.paginator = SeekPaginator(
-            Article.objects.all(), per_page=10, lookup_field="date_unique")
-
-    def test_prepare_order(self):
-        self.assertListEqual(self.paginator.prepare_order(), ["-date_unique", "-pk"])
-        self.paginator.lookup_field = "pk"
-        self.assertListEqual(self.paginator.prepare_order(), ["-pk", ])
-        self.paginator.lookup_field = "id"
-        self.assertListEqual(self.paginator.prepare_order(), ["-id", ])
-
-    def test_prepare_lookup(self):
-        lookup_f, lookup_e = self.paginator.prepare_lookup(value=1, pk=2)
-        self.assertDictEqual(lookup_f, {"date_unique__lte": 1, })
-        self.assertDictEqual(lookup_e, {"date_unique": 1, "pk__gte": 2})
-
-        self.paginator.lookup_field = "pk"
-        lookup_f, lookup_e = self.paginator.prepare_lookup(value=2, pk=2)
-        self.assertDictEqual(lookup_f, {"pk__lt": 2, })
-        self.assertIsNone(lookup_e)
-
-        self.paginator.lookup_field = "id"
-        lookup_f, lookup_e = self.paginator.prepare_lookup(value=2, pk=2)
-        self.assertDictEqual(lookup_f, {"id__lt": 2, })
-        self.assertIsNone(lookup_e)
-
-    def test_paginator(self):
+    def test_paginator_prev_desc(self):
         articles = Article.objects.all().order_by("-date_unique")
-
-        page_1 = self.paginator.page()
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page_2 = paginator.page(
+            value=list(articles)[20].date_unique,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_1 = paginator.page(
+            value=page_2[0].date_unique,
+            move_to=inf_paginator.PREV_PAGE)
         self.assertListEqual(list(page_1), list(articles[:10]))
 
-        page_2 = self.paginator.page(value=page_1[-1].date_unique, pk=page_1[-1].pk)
+    def test_paginator_next_desc(self):
+        articles = Article.objects.all().order_by("-date_unique")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page_1 = paginator.page(value=None)
+        self.assertListEqual(list(page_1), list(articles[:10]))
+        page_2 = paginator.page(value=page_1[-1].date_unique)
         self.assertListEqual(list(page_2), list(articles[10:20]))
-
-        page_3 = self.paginator.page(value=page_2[-1].date_unique, pk=page_2[-1].pk)
+        page_3 = paginator.page(value=page_2[-1].date_unique)
         self.assertListEqual(list(page_3), list(articles[20:]))
 
-    def test_first_page(self):
-        page = self.paginator.page()
-        self.assertTrue(page.has_next())
+    def test_paginator_prev_asc(self):
+        articles = Article.objects.all().order_by("date_unique")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="date_unique")
+        page_2 = paginator.page(
+            value=list(articles)[20].date_unique,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_1 = paginator.page(
+            value=page_2[0].date_unique,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page_1), list(articles[:10]))
 
-    def test_last_page(self):
-        articles = Article.objects.all().order_by("-date_unique")
-        next_to_last = list(articles)[-2]
+    def test_paginator_next_asc(self):
+        articles = Article.objects.all().order_by("date_unique")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="date_unique")
+        page_1 = paginator.page(value=None)
+        self.assertListEqual(list(page_1), list(articles[:10]))
+        page_2 = paginator.page(value=page_1[-1].date_unique)
+        self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_3 = paginator.page(value=page_2[-1].date_unique)
+        self.assertListEqual(list(page_3), list(articles[20:]))
 
-        page = self.paginator.page(value=next_to_last.date_unique, pk=next_to_last.pk)
-        self.assertFalse(page.has_next())
-
-    def test_lookup_not_unique(self):
-        self.paginator.lookup_field = "date"
+    def test_paginator_prev_desc_non_unique(self):
         articles = Article.objects.all().order_by("-date", "-pk")
-
-        page_1 = self.paginator.page()
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date")
+        page_2 = paginator.page(
+            value=list(articles)[20].date,
+            pk=list(articles)[20].pk,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_1 = paginator.page(
+            value=page_2[0].date,
+            pk=page_2[0].pk,
+            move_to=inf_paginator.PREV_PAGE)
         self.assertListEqual(list(page_1), list(articles[:10]))
 
-        page_2 = self.paginator.page(value=page_1[-1].date, pk=page_1[-1].pk)
+    def test_paginator_next_desc_non_unique(self):
+        articles = Article.objects.all().order_by("-date", "-pk")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date")
+        page_1 = paginator.page(value=None, pk=None)
+        self.assertListEqual(list(page_1), list(articles[:10]))
+        page_2 = paginator.page(value=page_1[-1].date, pk=page_1[-1].pk)
         self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_3 = paginator.page(value=page_2[-1].date, pk=page_2[-1].pk)
+        self.assertListEqual(list(page_3), list(articles[20:]))
 
-    def test_lookup_pk(self):
-        self.paginator.lookup_field = "pk"
-        articles = Article.objects.all().order_by("-pk")
-
-        page_1 = self.paginator.page()
+    def test_paginator_prev_asc_non_unique(self):
+        articles = Article.objects.all().order_by("-date", "-pk")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date")
+        page_2 = paginator.page(
+            value=list(articles)[20].date,
+            pk=list(articles)[20].pk,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_1 = paginator.page(
+            value=page_2[0].date,
+            pk=page_2[0].pk,
+            move_to=inf_paginator.PREV_PAGE)
         self.assertListEqual(list(page_1), list(articles[:10]))
 
-        page_2 = self.paginator.page(value=page_1[-1].pk, pk=page_1[-1].pk)
-        self.assertListEqual(list(page_2), list(articles[10:20]))
-
-    def test_lookup_id(self):
-        self.paginator.lookup_field = "id"
-        articles = Article.objects.all().order_by("-id")
-
-        page_1 = self.paginator.page()
+    def test_paginator_next_asc_non_unique(self):
+        articles = Article.objects.all().order_by("-date", "-pk")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date")
+        page_1 = paginator.page(value=None, pk=None)
         self.assertListEqual(list(page_1), list(articles[:10]))
-
-        page_2 = self.paginator.page(value=page_1[-1].id, pk=page_1[-1].id)
+        page_2 = paginator.page(value=page_1[-1].date, pk=page_1[-1].pk)
         self.assertListEqual(list(page_2), list(articles[10:20]))
+        page_3 = paginator.page(value=page_2[-1].date, pk=page_2[-1].pk)
+        self.assertListEqual(list(page_3), list(articles[20:]))
 
     def test_reverse_date_for_pk(self):
         """
@@ -106,24 +130,25 @@ class PaginatorTest(TestCase):
         we should still get the right results.
         """
         Article.objects.all().delete()
+        self.assertFalse(list(Article.objects.all()))
 
         # asc order date and desc order pk
         date = timezone.now()
-        dates = reversed([date + datetime.timedelta(seconds=seconds)
-                          for seconds in range(25)])
+        dates = reversed(
+            [date + datetime.timedelta(seconds=seconds)
+             for seconds in range(25)])
 
         for i, d in enumerate(dates):
             Article.objects.create(title="%s" % i, date=date, date_unique=d)
 
         articles = Article.objects.all().order_by("-date_unique")
-
-        page_1 = self.paginator.page()
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page_1 = paginator.page(value=None)
         self.assertListEqual(list(page_1), list(articles[:10]))
-
-        page_2 = self.paginator.page(value=page_1[-1].date_unique, pk=page_1[-1].pk)
+        page_2 = paginator.page(value=page_1[-1].date_unique, pk=page_1[-1].pk)
         self.assertListEqual(list(page_2), list(articles[10:20]))
-
-        page_3 = self.paginator.page(value=page_2[-1].date_unique, pk=page_2[-1].pk)
+        page_3 = paginator.page(value=page_2[-1].date_unique, pk=page_2[-1].pk)
         self.assertListEqual(list(page_3), list(articles[20:]))
 
 
@@ -136,34 +161,108 @@ class PageTest(TestCase):
             seconds = datetime.timedelta(seconds=i)
             Article.objects.create(title="%s" % i, date=date, date_unique=date + seconds)
 
-        self.paginator = SeekPaginator(
+    def test_next_page(self):
+        articles = Article.objects.all().order_by("date_unique")
+        paginator = SeekPaginator(
             Article.objects.all(), per_page=10, lookup_field="date_unique")
+        page = paginator.page(value=None)
+        self.assertListEqual(list(page), list(articles[:10]))
+        page = paginator.page(**page.next_page())
+        self.assertListEqual(list(page), list(articles[10:20]))
+        page = paginator.page(**page.next_page())
+        self.assertListEqual(list(page), list(articles[20:]))
 
-    def test_unimplemented(self):
-        page = self.paginator.page()
-        self.assertRaises(NotImplementedError, page.has_previous)
-        self.assertRaises(NotImplementedError, page.next_page_number)
-        self.assertRaises(NotImplementedError, page.previous_page_number)
-        self.assertRaises(NotImplementedError, page.start_index)
-        self.assertRaises(NotImplementedError, page.end_index)
-
-    def test_objects_left(self):
+    def test_prev_page(self):
         articles = Article.objects.all().order_by("-date_unique")
-        page = self.paginator.page()
-        self.assertEqual(page.objects_left, len(articles[self.paginator.per_page:]))
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(
+            value=list(articles)[20].date_unique,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page), list(articles[10:20]))
+        page = paginator.page(
+            **page.prev_page(),
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertListEqual(list(page), list(articles[:10]))
 
+    def test_next_objects_left(self):
+        articles = Article.objects.all().order_by("-date_unique")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(value=None)
+        self.assertEqual(
+            page.next_objects_left(),
+            len(articles[paginator.per_page:]))
         # last page
-        art = list(articles)[-self.paginator.per_page]
-        page_last = self.paginator.page(value=art.date_unique, pk=art.pk)
-        self.assertEqual(page_last.objects_left, 0)
+        page_last = paginator.page(
+            value=list(articles)[-paginator.per_page].date_unique)
+        self.assertEqual(page_last.next_objects_left(), 0)
 
-    def test_pages_left(self):
-        page = self.paginator.page()
-        self.assertEqual(page.pages_left, 2)
+    def test_prev_objects_left(self):
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(value=None)
+        self.assertEqual(page.prev_objects_left(), 0)
+        page = paginator.page(**page.next_page())
+        self.assertEqual(page.prev_objects_left(), paginator.per_page)
+        page = paginator.page(**page.next_page())
+        self.assertEqual(page.prev_objects_left(), paginator.per_page * 2)
 
-    def test_next_page_pk(self):
-        page = self.paginator.page()
-        self.assertEqual(page.next_page_pk(), page[-1].pk)
+    def test_next_pages_left(self):
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(value=None)
+        self.assertEqual(page.next_pages_left(), 2)
+        page = paginator.page(**page.next_page())
+        self.assertEqual(page.next_pages_left(), 1)
+        page = paginator.page(**page.next_page())
+        self.assertEqual(page.next_pages_left(), 0)
+
+    def test_prev_pages_left(self):
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(value=None)
+        self.assertEqual(page.prev_pages_left(), 0)
+        page = paginator.page(**page.next_page())
+        self.assertEqual(page.prev_pages_left(), 1)
+        page = paginator.page(**page.next_page())
+        self.assertEqual(page.prev_pages_left(), 2)
+
+    def test_has_next_page(self):
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="date_unique")
+        page = paginator.page(value=None)
+        self.assertTrue(page.has_next())
+        page = paginator.page(**page.next_page())
+        self.assertTrue(page.has_next())
+        page = paginator.page(**page.next_page())
+        self.assertFalse(page.has_next())
+
+    def test_has_prev_page(self):
+        articles = Article.objects.all().order_by("-date_unique")
+        paginator = SeekPaginator(
+            Article.objects.all(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(
+            value=list(articles)[20].date_unique,
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertTrue(page.has_previous())
+        page = paginator.page(
+            **page.prev_page(),
+            move_to=inf_paginator.PREV_PAGE)
+        self.assertFalse(page.has_previous())
+
+    def test_empty_first_page(self):
+        paginator = SeekPaginator(
+            Article.objects.none(), per_page=10, lookup_field="-date_unique")
+        page = paginator.page(value=None)
+        self.assertFalse(list(page))
+        self.assertFalse(page.has_next())
+        self.assertFalse(page.has_previous())
+        self.assertEqual(page.next_objects_left(), 0)
+        self.assertEqual(page.prev_objects_left(), 0)
+        self.assertEqual(page.next_pages_left(), 0)
+        self.assertEqual(page.next_page(), {})
+        self.assertEqual(page.prev_page(), {})
 
 
 class PaginatorViewTest(TestCase):
@@ -181,14 +280,18 @@ class PaginatorViewTest(TestCase):
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         res = json.loads(response.content.decode('utf-8'))
         articles = Article.objects.all().order_by("-date", "-pk")
-        self.assertEqual(res['articles'], [{u'title': a.title, } for a in articles[:20]])
+        self.assertEqual(
+            res['articles'],
+            [{u'title': a.title, } for a in articles[:20]])
 
     def test_page(self):
-        articles = Article.objects.all().order_by("-date", "-pk")
+        articles = list(Article.objects.all().order_by("-date", "-pk"))
         art = articles[20]
-
+        key = dict(value=art.date, pk=art.pk)
         response = self.client.get(
-            reverse('pagination-ajax', kwargs={'pk': str(art.pk), }),
+            reverse('pagination-ajax') + '?p={}'.format(to_page_key(**key)),
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         res = json.loads(response.content.decode('utf-8'))
-        self.assertEqual(res['articles'], [{u'title': a.title, } for a in articles[21:40]])
+        self.assertEqual(
+            res['articles'],
+            [{u'title': a.title, } for a in articles[21:]])